Extract chat discussion state
